refactor: split batch commands by responsibility
