+++ jvm/src/main/scala/io/github/apexdevtools/apexls/BatchCommand.scala
@@ -1,0 +1,32 @@
+/*
+ Copyright (c) 2026 Kevin Jones, All rights reserved.
+ Redistribution and use in source and binary forms, with or without
+ modification, are permitted provided that the following conditions
+ are met:
+ 1. Redistributions of source code must retain the above copyright
+    notice, this list of conditions and the following disclaimer.
+ 2. Redistributions in binary form must reproduce the above copyright
+    notice, this list of conditions and the following disclaimer in the
+    documentation and/or other materials provided with the distribution.
+ 3. The name of the author may not be used to endorse or promote products
+    derived from this software without specific prior written permission.
+ */
+
+package io.github.apexdevtools.apexls
+
+import com.nawforce.apexlink.api.Org
+
+private[apexls] final case class BatchContext(options: BatchOptions, org: Option[Org])
+
+private[apexls] trait BatchCommand {
+  type Result
+
+  def name: String
+  def requiresWorkspace: Boolean
+  def validate(args: Seq[String]): Either[BatchError, Unit] = {
+    if (args.isEmpty) Right(())
+    else Left(BatchError("INVALID_ARGUMENT", s"Unexpected argument '${args.head}'"))
+  }
+  def execute(context: BatchContext, args: Seq[String]): Either[BatchError, Result]
+  def writeResult(result: Result): ujson.Value
+}
